requirements.txt
requests
pymongo
python-dotenv
selenium
webdriver-manager
autopep8
